python/requirements.txt
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
